style: Customer 관련 클래스들 Customer.swift로 이동
