+++ BankManager.swift
@@ -6,37 +6,6 @@
 
 import Foundation
 
-class BankManager {
-    private var numberOfBanker: Int
-    private var waitingLine: OperationQueue
-    private var customerMaker: CustomerMaker
-    
-    init(numberOfBanker: Int) {
-        self.numberOfBanker = numberOfBanker
-        waitingLine = OperationQueue()
-        customerMaker = CustomerMaker()
-        waitingLine.maxConcurrentOperationCount = self.numberOfBanker
-    }
-    
-    func inputCustomersToWaitingLine() {
-        customerMaker.makeCustomers().forEach{ waitingLine.addOperation($0.bankTask) }
-        waitingLine.waitUntilAllOperationsAreFinished()
-    }
-}
-
-class Customer {
-    private var _bankTask: BankTask
-    private var waitingNumber: Int
-    var bankTask: BankTask {
-        return _bankTask
-    }
-    
-    init(waitingNumber: Int) {
-        self._bankTask = BankTask(taskTime: 0.7, waitingNumber)
-        self.waitingNumber = waitingNumber
-    }
-}
-
 class CustomerMaker {
     private var numberOfCustomer: Int
     
@@ -53,18 +22,20 @@ class CustomerMaker {
     }
 }
 
-class BankTask: Operation {
-    private var taskTime: Double
-    private var waitingNumber: Int
+class BankManager {
+    private var numberOfBanker: Int
+    private var waitingLine: OperationQueue
+    private var customerMaker: CustomerMaker
     
-    init(taskTime: Double, _ waitingNumber: Int) {
-        self.taskTime = taskTime
-        self.waitingNumber = waitingNumber
+    init(numberOfBanker: Int) {
+        self.numberOfBanker = numberOfBanker
+        waitingLine = OperationQueue()
+        customerMaker = CustomerMaker()
+        waitingLine.maxConcurrentOperationCount = self.numberOfBanker
     }
     
-    override func main() {
-        print(waitingNumber, "번 고객 업무 시작")
-        Thread.sleep(forTimeInterval: taskTime)
-        print(waitingNumber, "번 고객 업무 완료")
+    func inputCustomersToWaitingLine() {
+        customerMaker.makeCustomers().forEach{ waitingLine.addOperation($0.bankTask) }
+        waitingLine.waitUntilAllOperationsAreFinished()
     }
 }
